requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
slowapi==0.1.9
cryptography==41.0.7
pytest==7.4.3
python-dotenv==1.0.0
ccxt
slowapi